fix: Update shortcuts in onResume() callback instead of next launch
While on that also cleanup MainActivity & extract some UI related logic
